+++ source_code/src/leds.h
@@ -18,7 +18,7 @@ enum LEDS { LED_FRONT = 3,
             LED_CENTER = 6 };
 
 void init_rgb(void);
-void set_rgb(enum LEDS led, int r, int g, int b);
+void set_rgb(enum LEDS led, uint8_t r, uint8_t g, uint8_t b);
 void clear_rgb(void);
 void update_rgb(void);
 void manage_rgb(void);
